6.1.2 DI를 이용한 클래스의 분리 (트랜잭션 경계설정을 담당하는 코드를 외부로 빼내기 위해) - DI를 이용하여 트랜잭션의 경계설정 책임만 맡고, 비즈니스 로직은 다른 클래스에게 위임시킨다. 이를 통해서 비즈니스 로직을 담당하는 부분은 트랜잭션과 같은 기술적인 내용에는 전혀 신경 쓰지 않아도 된다. - UserService 인터페이스를 도입하여 비즈니스 로직만을 구현한 구현체(UserServiceImpl)를 만든다. - 트랜잭션 경계설정이라는 부가적인 작업이 부여된 UserServiceTx는 비즈니스 로직을 다른 UserService 인터페이스 구현체에게 위임한다. - 코드를 검증하기 위한 테스트의 경우, 어떤 클래스의 오브젝트인지 분명하게 알 필요가 있으므로 구현체를 직접 선언하여 테스트 빈을 주입받도록 한다.
